int n

+++ fusaco/2024_2025/2425S2Q1/main.cpp
@@ -8,7 +8,7 @@ using namespace std;
 using ll = long long;
 
 int main() {
-    ll n;
+    int n;
     cin >> n;
     vector<int> a(n), b(n) ;
